[FEATURE-13] Move files to FSD folders
