+++ src/Host/ClientApp/src/app/App.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react';
 import { Route, Routes } from 'react-router';
-import Layout from './components/Layout';
-import Home from './components/Home';
-import Counter from './components/Counter';
-import FetchData from './components/FetchData';
+import Layout from './Layout';
+import Home from '../pages/Home';
+import Counter from '../pages/Counter';
+import FetchData from '../pages/FetchData';
 
 import './custom.css'
