+++ src/Host/ClientApp/src/app/custom.css
