+++ src/Host/ClientApp/src/app/localhostApi.ts
