+++ src/Host/ClientApp/src/index.tsx
@@ -2,9 +2,9 @@ import * as React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
+import App from './app/App';
 import registerServiceWorker from './registerServiceWorker';
-import { setupStore } from './store';
+import { setupStore } from './app';
 
 // Получает экземпляр хранилища (store) для всего приложения
 const store = setupStore();
