+++ src/Host/ClientApp/src/widgets/NavMenu.tsx
